Move UserPinAnnotationView into its own file
and add doc comments
